split routing out of agent_config.c
agent_config.c reached 3343 lines against line-check's 2500 hard cap, so CI
could not go green. Move everything under the file's own "Routing" banner
into server/agent_route.c verbatim: role/persona matching, the route filters,
the cheapest-seat-with-capability router and the escalation helpers. Pure code
motion, no behaviour change; agent_config.c is left at 2437.

agent_endpoint_is_localish is the one helper both halves need — agent_config.c
to decide respond-tool injection, agent_route.c to classify a seat as local for
prefer_local. It becomes external via a narrow agent_config_internal.h rather
than widening the public agent_config.h surface.

Also move agent_tier_lint.c from CORE_SRCS to DATA_SRCS. CORE_SRCS feeds
KB_CORE_SRCS, and kb-target-isolation forbids agent_* sources in the KB
targets; the file is CLI/server price policy (cmd_doctor, cmd_agent,
server_agent) and was never needed by aimee-kb.
